requirements.txt
flask
jinja2
google-generativeai
python-dotenv 